+++ backend/reitera-backend/src/main/java/com/helderruiz/reitera_backend/modules/card/dto/CardResponseDTO.java
@@ -10,10 +10,10 @@ import java.util.Map;
 public record CardResponseDTO(
         Integer id,
         Integer deckId,
+        Integer noteId,
         String type,
         String question,
         Map<String, Object> answerJson,
-        String explanation,
         Integer state
 ) {
 }